Auto format

+++ Patches/ResearchRateMultiplicativeLabsPatch.cs
@@ -12,19 +12,15 @@ namespace ScienceScaling.Patches;
 static class ResearchRateMultiplicativeLabsPatch {
     [HarmonyPrefix]
     static bool Prefix(ResearchManager __instance, ResearchDefinition rd, ref float __result) {
-        if (!Plugin.Enabled.Value) {
-            return true;
-        }
+        if (!Plugin.Enabled.Value) { return true; }
 
         var s = Mathf.Clamp(__instance.company.BonusController.GetBonus(EBonus.ResearchProduction), 0.1f, 1f);
-        var num = (float)__instance.researchPoint1PerHour * Plugin.BaseRateMultiplier.Value / s;
-        if ((double)__instance.bonusFromObservatory > 0.0) {
-            num *= 2f;
-        }
+        var num = __instance.researchPoint1PerHour * Plugin.BaseRateMultiplier.Value / s;
+        if (__instance.bonusFromObservatory > 0.0) { num *= 2f; }
         var k = Plugin.LabContributionPerPoint.Value;
         var result = num * (1f + k * __instance.company.BonusController.GetBonusFromLab(rd));
         if (DevConsoleCommands.speedResearchMultiplayer.HasValue) {
-            result *= (float)DevConsoleCommands.speedResearchMultiplayer.Value;
+            result *= DevConsoleCommands.speedResearchMultiplayer.Value;
         }
         __result = result;
         return false;
